requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
attrs==26.1.0
certifi==2026.4.22
charset-normalizer==3.4.7
dateparser==1.4.0
frozenlist==1.8.0
idna==3.13
multidict==6.7.1
propcache==0.4.1
pycryptodome==3.23.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python_binance==1.0.36
pytz==2026.1.post1
regex==2026.4.4
requests==2.33.1
six==1.17.0
tzlocal==5.3.1
urllib3==2.6.3
websockets==16.0
yarl==1.23.0
